# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.APIError:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
  handlers.APIGenerateKeySetRequest:
    properties:
      name:
        example: production
        type: string
      secureBootEnroll:
        enum:
        - "off"
        - manual
        - if-safe
        - force
        type: string
    type: object
...
  handlers.APIRegistrationTokenResponse:
    properties:
      registrationToken:
        type: string
    type: object
...
  store.SecureBootKeySet:
    properties:
      createdAt:
        type: string
      id:
        type: string
      keysDir:
        type: string
      name:
        type: string
      secureBootEnroll:
        type: string
      tpmPcrKeyPath:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact:
    name: Kairos authors
    url: https://github.com/kairos-io/AuroraBoot
  description: |-
    HTTP API for the AuroraBoot Kairos node manager.

    Three security schemes are used across the API:
    * AdminBearer — Authorization: Bearer <admin-password>. Human/UI/automation callers.
    * NodeAPIKey — Authorization: Bearer <api-key>. Registered Kairos nodes calling their own endpoints. The api-key is returned from POST /api/v1/nodes/register.
    * RegistrationToken — shared secret carried inside the body of POST /api/v1/nodes/register.

    Two WebSocket endpoints also exist but are not described in this spec:
    * GET /api/v1/ws — agent command channel (auth via ?token=<api-key>).
    * GET /api/v1/ws/ui — UI live update channel (auth via ?token=<admin-password>).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: AuroraBoot API
  version: 0.1.0
paths:
...
  /api/v1/secureboot-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.SecureBootKeySet'
            type: array
      security:
      - AdminBearer: []
      summary: List SecureBoot key sets
      tags:
      - SecureBoot
  /api/v1/secureboot-keys/{id}:
    delete:
      description: Removes the database record. Key files on disk are left in place.
      parameters:
      - description: Key set ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - AdminBearer: []
      summary: Delete a key set record
      tags:
      - SecureBoot
  /api/v1/secureboot-keys/{id}/export:
    get:
      description: Streams a gzipped tar with manifest.json at the root and every
        key file under keys/. File permissions preserved so private keys remain 0600.
      parameters:
      - description: Key set ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/gzip
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - AdminBearer: []
      summary: Export a key set as a portable tar.gz
      tags:
      - SecureBoot
  /api/v1/secureboot-keys/generate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Key set name + enroll mode
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.APIGenerateKeySetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.SecureBootKeySet'
      security:
      - AdminBearer: []
      summary: Generate a SecureBoot key set
      tags:
      - SecureBoot
  /api/v1/secureboot-keys/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Override the manifest name to avoid a collision
        in: query
        name: name
        type: string
      - description: tar.gz produced by the export endpoint
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.SecureBootKeySet'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.APIError'
      security:
      - AdminBearer: []
      summary: Import a key set exported from another instance
      tags:
      - SecureBoot
  /api/v1/settings/registration-token:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIRegistrationTokenResponse'
      security:
      - AdminBearer: []
      summary: Read the current registration token
      tags:
      - Settings
  /api/v1/settings/registration-token/rotate:
    post:
      description: Generates and persists a new token. Already-registered nodes keep
        working because they authenticate with their api-key.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIRegistrationTokenResponse'
      security:
      - AdminBearer: []
      summary: Rotate the registration token
      tags:
      - Settings
securityDefinitions:
  AdminBearer:
    description: Supply as "Bearer <admin-password>".
    in: header
    name: Authorization
    type: apiKey
  NodeAPIKey:
    description: Supply as "Bearer <api-key>" returned from POST /api/v1/nodes/register.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"